+++ src/PYTHON/pair_python.cpp
@@ -26,7 +26,7 @@
 #include "memory.h"
 #include "update.h"
 #include "neigh_list.h"
-#include "python.h"
+#include "lmppython.h"
 #include "error.h"
 #include "python_compat.h"
